Хранимая процедура выполняется. Начала программировать авторизацию.

+++ Finance/Model/Entities/ExpensesPeriod.cs
@@ -10,7 +10,7 @@ namespace Finance.Model.Entities
     {
         public string Name { get; set; }
         public DateTime Date { get; set; }
-        public float Sum { get; set; }
+        public double Sum { get; set; }
         public string Category { get; set; }
         public string Necessity { get; set; }
     }
